chore: Rename key_test.ts to test.ts
